% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.helpers.R
\name{add.attribute}
\alias{add.attribute}
\title{Function to set a vertex  or edge attribute of a \code{network} or \code{graph} object}
\usage{
add.attribute(x, where = c("edge", "vertex"), attr_name, value, which = NULL)
}
\arguments{
\item{x}{The representation of the network as a \code{network} or \code{graph} object}

\item{where}{What network element does the attribute refer to. Either \code{edge}/\code{tie} or \code{vertex}/\code{node}.}

\item{attr_name}{Name of the attribute to set}

\item{value}{of the attribute to set}

\item{which}{A subset of elements on which the attribute should be applied. Defaults to all the vertexes/nodes.}
}
\value{
A \code{network} or \code{graph} object with the desired attribute
}
\description{
Function to set a vertex  or edge attribute of a \code{network} or \code{graph} object
}
\author{
\enc{Telarico, Fabio Ashtar}{Fabio Ashtar Telarico}
}
\keyword{internal}
