% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04.FF_methods.R
\name{as.firm,financial_matrix-method}
\alias{as.firm,financial_matrix-method}
\title{Coerce a \code{financial_matrix} object into a list of \code{firm} objects}
\usage{
\S4method{as.firm}{financial_matrix}(x, ...)
}
\arguments{
\item{x}{The \code{financial_matrix} object to coerce}

\item{...}{Optional arguments}
}
\value{
A (usually named) list of \code{firm} objects the length of which equals the number of rows and columns of the provided \code{financial_matrix}
}
\description{
`as.firm` method for an object of class \code{financial_matrix}
}
\author{
\enc{Telarico, Fabio Ashtar}{Fabio Ashtar Telarico}
}
