% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.firm_class.R
\name{as.firm}
\alias{as.firm}
\title{Coerce into (a list of) \code{firm} object(s)}
\usage{
as.firm(x, ...)
}
\arguments{
\item{x}{The object to coerce}

\item{...}{Arguments passed to class-specific methods}
}
\value{
An object of class \code{firm} or a (usually named) list of them, depending on the available method for the object being coerced.
}
\description{
Generic function to coerce other other classes into the S4 class \code{firm} representing a firm (legal person)
}
\author{
\enc{Telarico, Fabio Ashtar}{Fabio Ashtar Telarico}
}
