% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06.4.Bonacich.R
\name{bonacich-internal}
\alias{bonacich-internal}
\alias{bonacich.sparse}
\alias{bonacich.dense}
\title{Function to compute Bonacich centrality efficiently on sparse and dense matrices}
\usage{
bonacich.sparse(mat, exo = 1, alpha = 1, tol = 1e-07)

bonacich.dense(mat, exo = 1, alpha = 1, tol = 1e-07)
}
\arguments{
\item{mat}{A \code{matrix} object.}

\item{exo}{A value for the exogenous parameter of the Bonacich centrality. Defaults to \code{1}.}

\item{alpha}{A value for the alpha parameter of the Bonacich centrality. Defaults to a function of the matrix's largest eigenvalue.}

\item{tol}{A value for the tolerance of the Bonacich centrality. Defaults to \code{1e-07}.}
}
\value{
A scalar vector with the Bonacich centrality of each unit.
}
\description{
Called upon by \code{\link{bonacich}}.
}
\author{
\enc{Telarico, Fabio Ashtar}{Fabio Ashtar Telarico}
}
\keyword{internal}
