% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06.3.cfa.R
\name{cfa}
\alias{cfa}
\title{Perform cascade failure analysis}
\usage{
cfa(
  ...,
  ordering = "tot",
  custom.order = NULL,
  decreasing = TRUE,
  Rcpp = ifelse(requireNamespace("Rcpp", quietly = TRUE), yes = TRUE, no = FALSE)
)
}
\arguments{
\item{...}{Firm-Firm network in one of the following classes:
\itemize{
 \item{\code{financial_matrix} produced by \code{FF} and family;}
 \item{\code{network_financial} or \code{network} if the relevant package is installed;}
 \item{\code{igraph_financial} or \code{igraph} if the  relevant package is installed.}
}}

\item{ordering}{In what order to remove the firms, the completing ordering is always returned as part of the result. Take the following values:
\itemize{
 \item{\code{random} - Random order, corresponds to percolation theory}
 \item{\code{out} - By out-degree}
 \item{\code{in} - By in-degree}
 \item{\code{tot} - By sum of in- and out-degre (default)}
 \item{\code{custom} - Customised ordering via \code{custom.order}}
}}

\item{custom.order}{Order in which to remove the firms. If \code{ordering} is not 'custom', it is ignored. Defaults to \code{NULL}.}

\item{decreasing}{Logical, defaults to \code{TRUE}. Only evaluated if \code{ordering} is neither 'custom' nor 'random'.
\itemize{
 \item{if \code{TRUE} - the ordering is by decreasing degree and higher in-/out-/total-degree firms are removed first}
 \item{if \code{FALSE} - the ordering is by increasing degree and higher in-/out-/total-degree firms are removed last}
}}

\item{Rcpp}{Whether to use the \code{C++} or native-\code{R} version of the search algorithm. Defaults to \code{TRUE} if the package \code{Rcpp} is installed.}
}
\value{
A \code{data.frame} with one row for the result of the CFA after each node is removed. The columns report:
\itemize{
 \item{\code{l_scc} - Size of the largest strongly connected component}
 \item{\code{rem_id} - ID of the firm removed}
 \item{\code{rem_pos} - Position of the firm removed (row/column number)}
 \item{\code{n_scc} - Number of strongly connected components}
 \item{\code{n_rem} - Number of firms removed}
 \item{\code{n_left} - Number of firms left}
}
}
\description{
Cascade failure analysis (CFA) involves understanding how failures in one
part of the network might cascade to other parts. Networks capable of
isolating such failures or minimizing their effects demonstrate higher
robustness.
}
\examples{
# Create a matrix
mat <- matrix(c(
    0, 1, 0, 1, 0, 1, 0, 0,
    0, 0, 1, 0, 0, 0, 0, 0,
    1, 0, 0, 0, 0, 0, 0, 0,
    0, 1, 1, 0, 1, 0, 0, 0,
    0, 0, 0, 1, 0, 1, 0, 0,
    0, 0, 1, 0, 0, 0, 1, 0,
   0, 0, 0, 0, 0, 1, 0, 0,
    0, 0, 0, 0, 1, 0, 1, 1
  ),ncol = 8, byrow = TRUE)
# Add rownames
rownames(mat) <- paste0("Firm", LETTERS[1:ncol(mat)])

# Create a FF matrix
mat <- methods::new('financial_matrix',
                    M = mat,
                    relation = c('own'),
                    legal_form = c('JSC'),
                    sector = c('A.01'),
                    revenues = c(NA),
                    capitalisation = c(NA),
                    currency = c('USD'))

# Notice the differnce between:
# a. CFA with ordering by in-degree (decreasing)
# b. CFA with ordering by in-degree (increasing)
cfa(mat, ordering = 'in', decreasing = FALSE)
# But ordering by increasing (decreasing) in-degree is the
# same as ordering by decreasing (increasing) out-degree and
# vice versa!
cfa(mat, ordering = 'out', decreasing = FALSE) # By out-degree (increasing)

}
\references{
Elliott, Matthew, Benjamin Golub, and Matthew O. Jackson. ‘Financial Networks and Contagion’. American Economic Review 104, no. 10 (1 October 2014): 3115–53. \doi{10.1257/aer.104.10.3115}.
}
\author{
\enc{Telarico, Fabio Ashtar}{Fabio Ashtar Telarico}
}
