% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.helpers.R
\name{check.correct.grey}
\alias{check.correct.grey}
\title{Function to check whether the provided colour is actually a shade of grey and correct it if necessary}
\usage{
check.correct.grey(hex)
}
\arguments{
\item{hex}{The RGB colour to check}
}
\value{
A valid grey colour in RGB format
}
\description{
Function to check whether the provided colour is actually a shade of grey and correct it if necessary
}
\author{
\enc{Telarico, Fabio Ashtar}{Fabio Ashtar Telarico}
}
\keyword{internal}
