% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.helpers.R
\name{.set_names}
\alias{.set_names}
\title{Function to return an object after assigning new names}
\usage{
.set_names(x, names, where = c("col", "row", "attr"))
}
\arguments{
\item{x}{Object on which to operate}

\item{names}{New names}

\item{where}{What to change:
\itemize{
 \item{\code{col} - Column names}
 \item{\code{row} - Row names}
 \item{\code{names} - Names attribute}
}}
}
\value{
The original object, with new names
}
\description{
Combines \code{magrittr::set_colnames}, \code{magrittr::set_rownames}, \code{magrittr::set_names}
}
\author{
\enc{Telarico, Fabio Ashtar}{Fabio Ashtar Telarico}
}
\keyword{internal}
