% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.helpers.R
\name{extract.matrix}
\alias{extract.matrix}
\title{Function to extract the bare matrix from several complex objects}
\usage{
extract.matrix(x, .except.igraph = FALSE)
}
\arguments{
\item{x}{The representation of the network as an object of class:
\itemize{
 \item{\code{financial_matrix} produced by \code{FF} and family;}
 \item{\code{network_financial} or \code{network} if the relevant package is installed;}
 \item{\code{igraph_financial} or \code{igraph} if the  relevant package is installed.}
}}
}
\value{
A \code{matrix} or an object from the package \code{Matrix}
}
\description{
Function to extract the bare matrix from several complex objects
}
\seealso{
Other Internal matrix (de)constructrs: 
\code{\link{get.matrix.adjacency}()},
\code{\link{get.matrix.adjacency.both}()},
\code{\link{load.Matrix}()}
}
\author{
\enc{Telarico, Fabio Ashtar}{Fabio Ashtar Telarico}
}
\concept{Internal matrix (de)constructrs}
\keyword{internal}
