% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04.FF_methods.R
\name{FF-basic-methods}
\alias{FF-basic-methods}
\alias{rownames,financial_matrix-method}
\alias{colnames,financial_matrix-method}
\title{Basic methods for objects of class \code{financial_matrix}}
\usage{
\S4method{rownames}{financial_matrix}(x, do.NULL = TRUE, prefix = "row")

\S4method{colnames}{financial_matrix}(x, do.NULL = TRUE, prefix = "row")
}
\arguments{
\item{x}{The \code{financial_matrix} object to operate on}

\item{do.NULL}{Whether to use \code{NULL} names. Defaults to \code{FALSE}}

\item{prefix}{Prefix for created names (if \code{do.NULL} is \code{FALSE} and names are \code{NULL})}
}
\value{
A character vector of length equal to the number of rows (or columns) in the \code{financial_matrix} corresponding to the names of the rows (or columns)
}
\description{
Basic methods for objects of class \code{financial_matrix}
}
\details{
Mind that usually the rows and columns are named after the firm's tickers.
}
\author{
\enc{Telarico, Fabio Ashtar}{Fabio Ashtar Telarico}
}
