% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04.FF_methods.R
\name{FF-math-methods}
\alias{FF-math-methods}
\alias{isSymmetric,financial_matrix-method}
\alias{summary,financial_matrix-method}
\title{Mathematical methods for \code{financial_matrix} objects}
\usage{
\S4method{isSymmetric}{financial_matrix}(object, ...)

\S4method{summary}{financial_matrix}(object, ...)
}
\arguments{
\item{object}{The \code{financial_matrix} object to operate on}

\item{...}{Arguments passed to the relevant \code{matrix} method}
}
\value{
\itemize{
 \item \code{isSymmetric}: a boolean, \code{TRUE} if the matrix is symmetric, \code{FALSE} otherwise;
 \item \code{summary}: a list of length equal to the number of numeric attributes possed by the \code{financial_matrix} (maximum three, the matrix itself, revenues, and capitalisation) assumed as measured on the same scale and denominated in the same currency). Each element of the list of class \code{c('summaryDefault', 'table')} which has specialized \link[base]{format} and \link[base]{print} methods
}
}
\description{
\code{isSymmetric} checks only the matrix-like part
\code{summary} operates on all numeric attributes and the matrix-like part
}
\details{
Mathematical methods for \code{financial_matrix} objects
}
\author{
\enc{Telarico, Fabio Ashtar}{Fabio Ashtar Telarico}
}
