% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04.FF_methods.R
\name{FF-nrow-ncol}
\alias{FF-nrow-ncol}
\alias{ncol,financial_matrix-method}
\alias{nrow,financial_matrix-method}
\title{Number of rows/columns in a \code{financial_matrix} object}
\usage{
\S4method{ncol}{financial_matrix}(x)

\S4method{nrow}{financial_matrix}(x)
}
\arguments{
\item{x}{The \code{financial_matrix} object to operate on}
}
\value{
A single numeric, the number of rows (columns) in the matrix. It also prints a message to the console if any of the object's other attributes (e.g., capitalisation) is not conformed to the matrix's dimensions
}
\description{
Unlike most other methods (i.e., \code{duplicated}, \code{isSymmetric}, \code{summary}, \code{rownames}, and \code{colnames}), these methods act on both the matrix-like and the other components of a \code{financial_matrix} object.
}
\details{
Checks if the length of the names matches that of the other attributes that are not \code{NA} or structurally of unitary length (i.e., the slots \code{M} and \code{relation}).
}
\author{
\enc{Telarico, Fabio Ashtar}{Fabio Ashtar Telarico}
}
