% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04.FF_methods.R
\name{FF-subset-method}
\alias{FF-subset-method}
\alias{subset,financial_matrix-method}
\title{Method to subset a \code{financial_matrix}}
\usage{
\S4method{subset}{financial_matrix}(x, ...)
}
\arguments{
\item{x}{The \code{financial_matrix} object to operate on}

\item{...}{Arguments passed to the relevant \code{matrix} method}
}
\value{
A \code{financial_matrix} object, subsetted to the desired firms
}
\description{
Subsets all components of a \code{financial_matrix} object
}
\author{
\enc{Telarico, Fabio Ashtar}{Fabio Ashtar Telarico}
}
