% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04.FF_class.R
\docType{class}
\name{financial_matrix}
\alias{financial_matrix}
\title{An S4 class for the adjacency matrix produced by \code{\link{FF}} and related functions to represent the relations between firms (legal person)}
\description{
An S4 class for the adjacency matrix produced by \code{\link{FF}} and related functions to represent the relations between firms (legal person)
}
\section{Slots}{

\describe{
\item{\code{M}}{Adjacency matrix}

\item{\code{relation}}{Relations represented by the matrix}

\item{\code{legal_form}}{Legal form of the firms}

\item{\code{sector}}{Sector in which the firms operate}

\item{\code{revenues}}{Yearly revenues}

\item{\code{capitalisation}}{Firms' capitalisation}

\item{\code{currency}}{Currency in which the capitalisation and revenues are expressed}
}}

\keyword{internal}
