% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05.0.network&graph_class.R
\docType{class}
\name{financial_net}
\alias{financial_net}
\title{A virtual class for relational data extending the packages \href{https://statnet.org/}{network}and \href{https://igraph.org/}{igraph}}
\description{
This is the un-exported \link[methods:BasicClasses]{virtual} base for the S4 classes \code{igraph_financial} and \code{network_financial} to represent the relations between firms (legal person)
}
\section{Slots}{

\describe{
\item{\code{relation}}{Relations represented by the matrix}

\item{\code{vertex.size}}{Attribute used to determine the vertexes' size}

\item{\code{vertex.color}}{Attribute used to determine the vertexes' colour}

\item{\code{edge.width}}{Whether the edges' width reflect tie weight}

\item{\code{edge.greyscale}}{Whether the edges' color reflects tie weight}
}}

\keyword{internal}
