% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinNet-package.R
\docType{package}
\name{FinNet}
\alias{FinNet-package}
\alias{FinNet}
\title{Quickly Build and Manipulate Financial Networks}
\description{
Providing classes, methods, and functions to deal with financial networks.
Users can easily store information about both physical and legal persons
by using pre-made classes that are studied for integration with scraping
packages such as 'rvest' and 'RSelenium'.
Moreover, the package assists in creating various types of financial networks
depending on the type of relation between its units depending on the relation
under scrutiny (ownership, board interlocks, etc.), the desired tie type
(valued or binary), and renders them in the most common formats (adjacency
matrix, incidence matrix, edge list, 'igraph', 'network').
}
\seealso{
Useful links:
\itemize{
  \item \url{https://fatelarico.github.io/FinNet.html}
  \item Report bugs at \url{https://github.com/FATelarico/FinNet/issues}
}

}
\author{
\enc{Telarico, Fabio Ashtar}{Fabio Ashtar Telarico}
}
\keyword{internal}
