% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.firm_class.R
\docType{class}
\name{firm}
\alias{firm}
\title{A firm}
\description{
An S4 class to representing a firm (legal person) as a financial entity
}
\details{
This S4 class stores all information on a firm created using \link{register.firm} or \link{find.firm}.
The firm's property can be accessed using the relative function (see \link{query.firm})
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of the firm}

\item{\code{id}}{Provide an ID code for the firm.}

\item{\code{legal_form}}{Legal form of the firm}

\item{\code{sector}}{Sector in which the firm operates}

\item{\code{sector_classif}}{Activity sector classification (if any)}

\item{\code{revenues}}{Yearly revenues}

\item{\code{capitalisation}}{Firm's capitalisation}

\item{\code{management}}{Names of the members of the board}

\item{\code{ownership}}{Names of the owner(s)}

\item{\code{shares}}{Share owned by (each of) the owner(s)}

\item{\code{currency}}{Currency in which the capitalisation and revenues are expressed}
}}

\keyword{internal}
