% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.data.R
\docType{data}
\name{firms_BKB}
\alias{firms_BKB}
\title{Complete Berkshire Hathaway Portfolio}
\format{
Three objects of class \code{firm}.
}
\source{
- Divine, John. “The Complete Berkshire Hathaway Portfolio.” FInancial data. U.S. News & World Report, May 17, 2023. <https://money.usnews.com/investing/stock-market-news/articles/the-complete-berkshire-hathaway-portfolio>.
- ICE Data Services. “Nasdaq Stock Exchange & Dow Jones Indexes.” Financial data, May 21, 2023, <https://finance.yahoo.com/lookup/>.
}
\usage{
data('firms_BKB')
}
\description{
Data on Apple (AAPL), General Motors (GM), and British American Tobacco (BTI) extracted from Yahoo! Finance (on May 20, 2023) and formatted a \code{firm} objects.
}
\keyword{datasets}
