% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.data.R
\docType{data}
\name{firms_US}
\alias{firms_US}
\title{Three US firms}
\format{
Three objects of class \code{firm}.
}
\source{
ICE Data Services. “Nasdaq Stock Exchange & Dow Jones Indexes.” Financial data, May 21, 2023, <https://finance.yahoo.com/lookup/>
}
\usage{
data('firms_US')
}
\description{
Data on Apple (AAPL), General Motors (GM), and British American Tobacco (BTI) extracted from Yahoo! Finance (on May 20, 2023) and formatted a \code{firm} objects.
}
\keyword{datasets}
