% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.FOandFM.R
\name{FM}
\alias{FM}
\title{Function to create a (necessarily binary) firm-manager (FM) matrix}
\usage{
FM(..., id_as_firm_name = NULL, Matrix = NULL)
}
\arguments{
\item{...}{Either multiple objects of class \code{firm} or a list of such objects}

\item{id_as_firm_name}{Whether to use the ticker as the firm's name. Defaults to \code{TRUE} if all firms' id is neither \code{NULL} nor \code{NA}.}

\item{Matrix}{Whether to use the \href{https://cran.r-project.org/package=Matrix}{\code{Matrix} package}. Defaults to \code{TRUE} when there are more than 10,000 combinations and the package is installed.}
}
\value{
A matrix object of class \code{financial_matrix} (possibly using the \href{https://cran.r-project.org/package=Matrix}{\code{Matrix} package}) in which:
\describe{
 \item{the rows}{Represent firms;}
 \item{the columns}{Represent managers (usually physical persons).}
}
}
\description{
Function to create a (necessarily binary) firm-manager (FM) matrix
}
\examples{

# Create the FM matrix of Berkshire Hathaway's holdings
\donttest{
data('firms_BKB')
  FM <- FM(firms_BKB)
}

}
\seealso{
\link{FO.binary} \link{FO.naive} \link{FO.norm}
}
\author{
\enc{Telarico, Fabio Ashtar}{Fabio Ashtar Telarico}
}
