% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.helpers.R
\name{get.binarary.values}
\alias{get.binarary.values}
\title{Function to compute the binary values of the FM or FO matrix for multiple \code{firm} objects}
\usage{
get.binarary.values(firms, which, cols = NULL)
}
\arguments{
\item{firms}{List of objects on which to operate}

\item{which}{Whether to use ownership or management to construct the matrix}

\item{cols}{Possible values assumed by the enquired variable (determined autmotically if not provided)}
}
\value{
The values to be filled in the binary FO or FM matrix
}
\description{
Function to compute the binary values of the FM or FO matrix for multiple \code{firm} objects
}
\author{
\enc{Telarico, Fabio Ashtar}{Fabio Ashtar Telarico}
}
\keyword{internal}
