% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05.1.graph_methods.R
\name{igraph_E_iterators}
\alias{igraph_E_iterators}
\alias{E,igraph-method}
\alias{E,igraph_financial-method}
\alias{ecount,igraph_financial-method}
\alias{ecount,igraph-method}
\alias{gsize,igraph_financial-method}
\alias{gsize,igraph-method}
\title{igraph edge iterators for \code{igraph_financial} objects}
\usage{
\S4method{E}{igraph_financial}(x, ...)

\S4method{E}{igraph}(x, ...)

\S4method{ecount}{igraph_financial}(x, ...)

\S4method{ecount}{igraph}(x, ...)

\S4method{gsize}{igraph_financial}(x, ...)

\S4method{gsize}{igraph}(x, ...)
}
\arguments{
\item{x}{The \code{igraph_financial} object}

\item{...}{Other parameters passed to the corresponding method and/or \code{igraph} functions (see Details).}
}
\value{
The same result for both \code{igraph} and \code{igraph_financial} objects
\itemize{
 \item \code{E}: An edge sequence of the graph
 \item \code{ecount} and \code{gsize}: Number of edges, numeric scalar
}
}
\description{
Methods to extend igraph edge iterators and functions to \code{igraph_financial} objects
}
\author{
\enc{Telarico, Fabio Ashtar}{Fabio Ashtar Telarico}
}
