% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05.0.network&graph_class.R
\docType{class}
\name{igraph_financial}
\alias{igraph_financial}
\alias{igraph_financial-class}
\title{An S4 class for relational data extending the package \href{https://igraph.org/}{igraph}}
\description{
An S4 class for the network objects produced by the \code{\link{FF.graph}} and \code{\link{FF.graph.custom}} to represent the relations between firms (legal person)
}
\section{Slots}{

\describe{
\item{\code{data}}{The representation of the network as a \href{https://rdrr.io/cran/igraph/man/make_graph.html}{igraph} object}
}}

