% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.helpers.R
\name{listing}
\alias{listing}
\title{Function to list multiple objects passed as \code{...}}
\usage{
listing(..., naming = TRUE)
}
\arguments{
\item{...}{Objects on which to operate}

\item{naming}{Logical | Whether to name the list after the symbols of the provided objects. Defaults to \code{TRUE}}
}
\value{
A (possibly named) list of the provided objects.
}
\description{
Function to list multiple objects passed as \code{...}
}
\author{
\enc{Telarico, Fabio Ashtar}{Fabio Ashtar Telarico}
}
\keyword{internal}
