% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04.FF_class.R
\docType{class}
\name{Matrix&matrix}
\alias{Matrix&matrix}
\title{A (virtual) class union for matrix-like objects to be included in \code{financial_matrix}}
\description{
This is the un-exported basis for the S4 classes \code{financial_matrix}
}
\keyword{internal}
