% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05.0.network&graph_class.R
\docType{class}
\name{network_financial}
\alias{network_financial}
\alias{network_financial-class}
\title{An S4 class for relational data extending the package \href{https://statnet.org/}{network}}
\description{
An S4 class for the network objects produced by the \code{\link{FF.net}} and \code{\link{FF.net.custom}} functions to represent the relations between firms (legal person)
}
\section{Slots}{

\describe{
\item{\code{data}}{The representation of the network as a \href{https://rdrr.io/cran/network/man/network.html}{network} object}
}}

