% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05.1.network_methods.R
\name{network_operators}
\alias{network_operators}
\alias{edgecount,network_financial-method}
\alias{edgecount,network-method}
\alias{vertex.names,network-method}
\alias{vertex.names,network_financial-method}
\alias{network.size,network-method}
\alias{network.size,network_financial-method}
\title{Operators for \code{network_financial} objects}
\usage{
\S4method{edgecount}{network_financial}(x, ...)

\S4method{edgecount}{network}(x, ...)

\S4method{vertex.names}{network_financial}(x, ...)

\S4method{vertex.names}{network}(x, ...)

\S4method{network.size}{network_financial}(x, ...)

\S4method{network.size}{network}(x, ...)
}
\arguments{
\item{x}{The \code{network_financial} object}

\item{...}{Other parameters passed to the corresponding method and/or \code{network} functions (see Details).}
}
\value{
The same result for both \code{network} and \code{network_financial} objects
\itemize{
 \item \code{network.edgecount}: Number of edges, numeric scalar
 \item \code{network.vertex.names}: Names/Labels of the vertices, character vector
 \item \code{network::network.size}: Number of vertices, numeric scalar
}
}
\description{
Methods to extend operators from the package \code{network} to \code{network_financial} objects
}
\author{
\enc{Telarico, Fabio Ashtar}{Fabio Ashtar Telarico}
}
