% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.helpers.R
\name{palette}
\alias{palette}
\title{Function to create the default palette and ramp it up as needed}
\usage{
palette(length = 6)
}
\arguments{
\item{length}{Length of the desired palette}
}
\value{
A vector of RGB codes containing \code{length} colours
}
\description{
Function to create the default palette and ramp it up as needed
}
\author{
\enc{Telarico, Fabio Ashtar}{Fabio Ashtar Telarico}
}
\keyword{internal}
