% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05.1.network_methods.R
\name{plot_network-methods}
\alias{plot_network-methods}
\alias{plot_network,network_financial-method}
\alias{plot_network,network-method}
\title{network plotting for \code{igraph_financial} objects}
\usage{
\S4method{plot_network}{network_financial}(x, ...)

\S4method{plot_network}{network}(x, ...)
}
\arguments{
\item{x}{The \code{network_financial} object}

\item{...}{Other parameters passed to the corresponding method and/or \code{network} functions (see Details).}
}
\value{
For both \code{igraph} and \code{igraph_financial} objects, returns NULL invisibly. It is called to print the graph to any R device. (see method and \href{https://rdrr.io/cran/igraph/man/plot.igraph.html}{igraph::plot.igraph})
}
\description{
Methods to extend \code{network}'s plotting functions to \code{network_financial} objects
}
\author{
\enc{Telarico, Fabio Ashtar}{Fabio Ashtar Telarico}
}
