% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04.FF_methods.R
\name{print,financial_matrix-method}
\alias{print,financial_matrix-method}
\title{Print information on a \code{financial_matrix}}
\usage{
\S4method{print}{financial_matrix}(x)
}
\arguments{
\item{x}{The \code{financial_matrix} object to print}
}
\value{
No return value, called to print to the console information about the \code{financial_matrix} object:
\itemize{
  \item type of relation detail (common ownership, board interlocks, or both),
  \item Total capitalisation of the firms present (if known), broke down by currency (if more than one);
  \item Average total revenues (if known), broke down by currency (if more than one)
  \item A print-out of the matrix, up to the number of character given by \code{options('max.print')}.
}
}
\description{
Print method for the objects of class \code{financial_matrix}
}
\author{
\enc{Telarico, Fabio Ashtar}{Fabio Ashtar Telarico}
}
\keyword{internal}
