% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05.1.graph_methods.R
\name{print,igraph_financial-method}
\alias{print,igraph_financial-method}
\title{Print information on a \code{igraph_financial} object}
\usage{
\S4method{print}{igraph_financial}(x)
}
\arguments{
\item{x}{The object of class \code{igraph_financial} to be shown}
}
\value{
Nothing, called for printing to the console the following information:
\itemize{
 \item The first line contains the basic properties of the graph;
 \item The rest of the of the print-out contains the graph's attributes. After the attribute names, the kind of the attribute – graph (‘g’), vertex (‘v’) or edge (‘e’) – is denoted, and the type of the attribute as well, character (‘c’), numeric (‘n’), logical (‘l’), or other (‘x’).
}
}
\description{
Method to \code{print} objects of class \code{igraph_financial}
}
\details{
For more information see the relevant igraph function: \href{https://rdrr.io/cran/igraph/man/print.igraph.html}{print.igraph}
}
\author{
\enc{Telarico, Fabio Ashtar}{Fabio Ashtar Telarico}
}
\keyword{internal}
