% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05.1.network_methods.R
\name{print,network_financial-method}
\alias{print,network_financial-method}
\title{Print information on a \code{network_financial} object}
\usage{
\S4method{print}{network_financial}(x)
}
\arguments{
\item{x}{The object of class \code{network_financial} to be shown}
}
\value{
Prints the list of global attributes of the network as if \code{network::print.network()} had been called. See \href{https://rdrr.io/cran/network/man/network.html}{print.network} for more details.
}
\description{
Method to \code{print} objects of class \code{network_financial}
}
\author{
\enc{Telarico, Fabio Ashtar}{Fabio Ashtar Telarico}
}
\keyword{internal}
