% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04.FF_methods.R
\name{print_financial_matrix}
\alias{print_financial_matrix}
\title{Print information on a \code{financial_matrix}}
\usage{
print_financial_matrix(x)
}
\arguments{
\item{x}{The object of class \code{financial_matrix} to be shown}
}
\description{
Un-exported function behind the \code{print} methods for objects of class \code{financial_matrix}
}
\author{
\enc{Telarico, Fabio Ashtar}{Fabio Ashtar Telarico}
}
\keyword{internal}
