% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.helpers.R
\name{query.firms_switch}
\alias{query.firms_switch}
\title{Function to chose the right algorithm when querying one or more information over multiple \code{firm} objects}
\usage{
query.firms_switch(firm, which, naming = TRUE, unlisting = FALSE)
}
\arguments{
\item{firm}{List of objects on which to operate}

\item{naming}{Logical | Whether to name the result after \code{names(firms)}. Defaults to \code{TRUE}}

\item{unlisting}{Logical | Whether to un-list the result. Defaults to \code{FALSE}}
}
\value{
The queried information.
}
\description{
Function to chose the right algorithm when querying one or more information over multiple \code{firm} objects
}
\author{
\enc{Telarico, Fabio Ashtar}{Fabio Ashtar Telarico}
}
\keyword{internal}
