% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.helpers.R
\name{rescale.numeric}
\alias{rescale.numeric}
\title{Function to rescale numeric vectors}
\usage{
rescale.numeric(
  x,
  to = c(0, 1),
  from = range(x, na.rm = TRUE, finite = TRUE),
  ...
)
}
\arguments{
\item{x}{A vector to re-scale}

\item{to}{output range (numeric vector of length two)}

\item{from}{input range (vector of length two).  If not given, is
calculated from the range of \code{x}}
}
\value{
Re-scaled vector
}
\description{
Rescale continuous vector to have specified minimum and maximum
}
\details{
From: The package \href{https://scales.r-lib.org/}{\code{scales}}
}
\author{
\href{https://scales.r-lib.org/}{Hadley Wickham}
}
\keyword{internal}
