% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.helpers.R
\name{SCC2.prep}
\alias{SCC2.prep}
\title{Create list of edges for SCC2}
\usage{
SCC2.prep(x)
}
\arguments{
\item{x}{A square adjacency matrix representing the directed network.}
}
\value{
If there are at least two non-zero rows:
\itemize{
 \item{A \code{list} whose elements are \code{vector}s representing the
       ties between units as \code{c(starting_node, end_node)} for the
       function #' \code{SCC2()}. Otherwise,}
 \item{A \code{list} like those produced by \code{SCC2()} and \code{SCC()}.}
}

Note: Isolated nodes are not returned, but the object \code{x2} is modified
in place to store information on their number if the matrix has at least two
non-null rows.
}
\description{
Create list of edges for SCC2
}
\seealso{
[SCC2()]
}
\author{
\enc{Telarico, Fabio Ashtar}{Fabio Ashtar Telarico}
}
\keyword{internal}
