% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.firm_methods.R
\name{show,firm-method}
\alias{show,firm-method}
\title{Show information on a \code{firm}}
\usage{
\S4method{show}{firm}(object)
}
\arguments{
\item{object}{The \code{firm} object to show}
}
\value{
No return value, called to print to the console information about the \code{firm} object including:
\itemize{
  \item legal form (if any),
  \item revenues (if known),
  \item capitalisation (if known),
  \item number of members of the board/managers; and
  \item number of owners/shareholders.
}
}
\description{
Show method for the S4 class \code{firm} representing a firm (legal person)
}
\author{
\enc{Telarico, Fabio Ashtar}{Fabio Ashtar Telarico}
}
\keyword{internal}
