% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05.1.network_methods.R
\name{show,network_financial-method}
\alias{show,network_financial-method}
\title{Show information on a \code{network_financial}}
\usage{
\S4method{show}{network_financial}(object)
}
\arguments{
\item{object}{The \code{network_financial} object to show}
}
\description{
Show method for the objects of class \code{network_financial}
}
\author{
\enc{Telarico, Fabio Ashtar}{Fabio Ashtar Telarico}
}
\keyword{internal}
