% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.firm_methods.R
\name{show_firm}
\alias{show_firm}
\title{Show information on a class \code{firm} object}
\usage{
show_firm(object)
}
\arguments{
\item{object}{The object of class \code{firm} to be shown}
}
\value{
No return value, called to print to the console information about the \code{firm} object including:
\itemize{
  \item legal form (if any),
  \item revenues (if known),
  \item capitalisation (if known),
  \item number of members of the board/managers; and
  \item number of owners/shareholders.
}
}
\description{
Unexported function behind the \code{show} methods for objects of class \code{firm}
}
\author{
\enc{Telarico, Fabio Ashtar}{Fabio Ashtar Telarico}
}
\keyword{internal}
