% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05.1.network_methods.R
\name{show_network_financial}
\alias{show_network_financial}
\title{Show information on a \code{network_financial}}
\usage{
show_network_financial(x)
}
\arguments{
\item{x}{The object of class \code{network_financial} to be shown}
}
\description{
Un-exported function behind the \code{show} method for objects of class \code{network_financial}
}
\author{
\enc{Telarico, Fabio Ashtar}{Fabio Ashtar Telarico}
}
\keyword{internal}
