% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04.FF_methods.R
\name{subset_financial_matrix}
\alias{subset_financial_matrix}
\title{Subset a \code{financial_matrix} object}
\usage{
subset_financial_matrix(x, ...)
}
\arguments{
\item{x}{The object of class \code{financial_matrix} to be subset}

\item{...}{Additional arguments to the \code{matrix} method for \code{subset()} (see under 'See also')}
}
\description{
Un-exported function behind the \code{subset} method for objects of class \code{financial_matrix}
}
\author{
\enc{Telarico, Fabio Ashtar}{Fabio Ashtar Telarico}
}
\keyword{internal}
