% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05.1.network_methods.R
\name{show,network_financial-method}
\alias{show,network_financial-method}
\title{Show information on a \code{network_financial}}
\usage{
\S4method{show}{network_financial}(object)
}
\arguments{
\item{object}{The \code{network_financial} object to show}
}
\value{
Nothing, called for printing to the console the following information:
\itemize{
 \item The first line contains the basic properties of the graph;
 \item The rest of the of the first paragrah contains the graph's attributes. After the attribute names, the kind of the attribute – graph (‘g’), vertex (‘v’) or edge (‘e’) – is denoted, and the type of the attribute as well, character (‘c’), numeric (‘n’), logical (‘l’), or other (‘x’);
 \item The second paragraph decribes the type of relation represented in the network (common ownership, board interlocks, both);
 \item The third paragraph informs on whether the size of the vertexes represents an attribute (and, if yes, which one);
 \item The fourth paragraph informs on whether the colour of the vertexes represents an attribute (and, if yes, which one);
 \item The fifth paragraph informs on whether the width of the edges represents their strength;
 \item The sixth paragraph informs on whether the colour of the edges represents their strength.
}
}
\description{
Show method for the objects of class \code{network_financial}
}
\author{
\enc{Telarico, Fabio Ashtar}{Fabio Ashtar Telarico}
}
\keyword{internal}
